- review of the AI generated code - adjusted module structure - new way to export data - new Transformer concept
