+++ modules/flexibleSearch/core/src/sap/commerce/toolset/flexibleSearch/FlexibleSearchConstants.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,6 +19,7 @@
 package sap.commerce.toolset.flexibleSearch
 
 import com.intellij.codeInsight.completion.CompletionUtilCore
+import com.intellij.openapi.util.Key
 import com.intellij.psi.tree.IFileElementType
 import sap.commerce.toolset.flexibleSearch.psi.FlexibleSearchTypes
 
@@ -80,4 +81,13 @@ object FlexibleSearchConstants {
         FlexibleSearchTypes.WHEN,
         FlexibleSearchTypes.WHERE,
     )
+
+    object Transform {
+        val INCLUDE_TYPE_SYSTEM_UNIQUE = Key.create<Boolean>("flexibleSearch.transform.includeTypeSystemUnique")
+        val INCLUDE_DATA = Key.create<Boolean>("flexibleSearch.transform.includeData")
+    }
+
+    object ImpEx {
+        const val IMPEX_IGNORE = "<ignore>"
+    }
 }
