+++ modules/flexibleSearch/mcp/src/sap/commerce/toolset/flexibleSearch/mcp/dto/FlexibleSearchMcpResult.kt
@@ -16,41 +16,20 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-fun properties(key: String) = providers.gradleProperty(key)
+package sap.commerce.toolset.flexibleSearch.mcp.dto
 
-plugins {
-    id("org.jetbrains.intellij.platform.module")
-    alias(libs.plugins.kotlin) // Kotlin support
-}
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+import sap.commerce.toolset.flexibleSearch.exec.context.FlexibleSearchExecResult
 
-sourceSets {
-    main {
-        java.srcDirs("src", "gen")
-        resources.srcDirs("resources")
-    }
-    test {
-        java.srcDirs("tests")
-    }
-}
-
-idea {
-    module {
-        generatedSourceDirs.add(file("gen"))
-    }
-}
-
-dependencies {
-    implementation(project(":shared-core"))
-    implementation(project(":typeSystem-core"))
-    implementation(project(":project-core"))
-
-    intellijPlatform {
-        intellijIdea(properties("intellij.version")) {
-            useInstaller = false
-        }
-
-        bundledPlugins(
-            "com.intellij.java",
-        )
-    }
-}
+@Serializable
+data class FlexibleSearchMcpResult(
+    val connection: String,
+    val success: Boolean,
+    val output: String? = null,
+    val description: String? = null,
+    val error: String? = null,
+    val errorDetail: String? = null,
+    @Transient
+    val rawResult: FlexibleSearchExecResult? = null
+)
