--- modules/flexibleSearch/mcp/src/sap/commerce/toolset/flexibleSearch/mcp/dto/FxSImpExResult.kt
@@ -1,33 +1,0 @@
-/*
- * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package sap.commerce.toolset.flexibleSearch.mcp.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FxSImpExResult(
-    val connection: String,
-    val success: Boolean,
-    val primaryType: String? = null,
-    val columnCount: Int? = null,
-    val rowCount: Int? = null,
-    val impex: String? = null,
-    val error: String? = null,
-    val errorDetail: String? = null,
-)
