+++ modules/flexibleSearch/transform/src/sap/commerce/toolset/flexibleSearch/transform/context/FxSColumn.kt
@@ -1,0 +1,38 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package sap.commerce.toolset.flexibleSearch.transform.context
+
+/**
+ * Analyzed representation of a FlexibleSearch SELECT column for ImpEx conversion.
+ *
+ * @param resultHeaderName  The name as returned in the HAC result headers (alias if present, otherwise column name).
+ * @param attributeName     The ImpEx attribute name (the actual attribute, not the alias).
+ * @param isPk              True if this column is the `pk` field — should be skipped in ImpEx output.
+ * @param isLocalized       True when using localized syntax `{col[en]}`.
+ * @param langCode          The language code when `isLocalized` is true (e.g., `en`).
+ * @param tableAlias        Table alias when using Y-column syntax `{alias:col}`.
+ */
+data class FxSColumn(
+    val resultHeaderName: String,
+    val attributeName: String,
+    val isPk: Boolean,
+    val isLocalized: Boolean = false,
+    val langCode: String? = null,
+    val tableAlias: String? = null,
+)
