+++ modules/flexibleSearch/ui/resources/META-INF/sap.commerce.toolset-flexibleSearch-ui.xml
@@ -49,7 +49,7 @@
         <action id="hybris.fxs.toggleParameters" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchToggleInEditorParametersAction"/>
         <action id="hybris.fxs.toggleInEditorResults" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchToggleInEditorResultsAction"/>
         <action id="hybris.fxs.copyToClipboard" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchCopyToClipboardAction"/>
-        <action id="hybris.fxs.exportToImpEx" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchExportToImpExAction" searchable="false"/>
+        <action id="hybris.fxs.export" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchExportAction"/>
         <action id="hybris.fxs.introduceBindParameters" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchIntroduceBindParametersAction"/>
         <action id="hybris.fxs.showRestrictions" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchShowRestrictionsAction"/>
         <action id="hybris.fxs.executionContextSettings" class="sap.commerce.toolset.flexibleSearch.actionSystem.FlexibleSearchExecutionContextSettingsAction"/>
@@ -60,7 +60,7 @@
             <reference ref="hybris.fxs.execute"/>
             <reference ref="hybris.fxs.openQuery"/>
             <reference ref="hybris.fxs.copyToClipboard"/>
-            <reference ref="hybris.fxs.exportToImpEx"/>
+            <reference ref="hybris.fxs.export"/>
             <reference ref="hybris.fxs.introduceBindParameters"/>
             <reference ref="hybris.fxs.showRestrictions"/>
             <separator/>
