+++ modules/impex/transform/resources/META-INF/sap.commerce.toolset-impex-transform.xml
@@ -15,11 +15,6 @@
   ~ You should have received a copy of the GNU Lesser General Public License
   ~ along with this program. If not, see <http://www.gnu.org/licenses/>.
   -->
-
 <idea-plugin>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.flexibleSearch.mcp.FlexibleSearchMcpToolset"/>
-    </extensions>
 
 </idea-plugin>
