+++ modules/shared/core/resources/i18n/HybrisBundle.properties
@@ -687,18 +687,20 @@ hybris.fxs.actions.open_query=Open FlexibleSearch Query in the Console
 hybris.fxs.actions.open_query.description=Copies and opens FlexibleSearch Query in the corresponding Console
 hybris.fxs.actions.copy_query_to_clipboard=Copy FlexibleSearch Query With Values
 hybris.fxs.actions.copy_query_to_clipboard.description=Copy Query with values to clipboard
-hybris.fxs.actions.export_to_impex=Export to ImpEx
-hybris.fxs.actions.export_to_impex.description=Convert FlexibleSearch results to an ImpEx INSERT_UPDATE block
-hybris.fxs.actions.export_to_impex.notification.title=ImpEx Exported to Clipboard
-hybris.fxs.actions.export_to_impex.notification.rows=rows
-hybris.fxs.actions.export_to_impex.notification.open_scratch=Open as Scratch File
-hybris.fxs.actions.export_to_impex.dialog.title=Export to ImpEx
-hybris.fxs.actions.export_to_impex.dialog.generate=Generate
-hybris.fxs.actions.export_to_impex.dialog.include_data=Include exported data
-hybris.fxs.actions.export_to_impex.dialog.include_data.has_result={0} rows from the last execution
-hybris.fxs.actions.export_to_impex.dialog.include_data.no_result=Execute the query first to include data
-hybris.fxs.actions.export_to_impex.dialog.clear_data=Clear data
-hybris.fxs.actions.export_to_impex.dialog.include_type_unique=Include all unique attributes from the type definition
+hybris.fxs.actions.export=Export FlexibleSearch
+hybris.fxs.actions.export.description=Convert FlexibleSearch and its results into a desired format
+hybris.fxs.actions.export.notification.title=FlexibleSearch Exported
+hybris.fxs.actions.export.notification.rows={0} rows
+hybris.fxs.actions.export.notification.open_scratch=Open as Scratch File
+hybris.fxs.actions.export.notification.copy_to_clipboard=Copy to Clipboard
+hybris.fxs.actions.export.dialog.title=Export FlexibleSearch
+hybris.fxs.actions.export.dialog.export=Export
+hybris.fxs.actions.export.dialog.include_data=Include exported data rows
+hybris.fxs.actions.export.dialog.include_data.has_result={0} rows from the last execution
+hybris.fxs.actions.export.dialog.include_data.no_result=Execute the query first to include data
+hybris.fxs.actions.export.dialog.clear_data=Clear Data
+hybris.fxs.actions.export.dialog.include_type_unique=Include all unique attributes from the type definition
+hybris.fxs.actions.export.dialog.transformer=Convert to:
 hybris.fxs.actions.introduce_bind_parameters=Introduce Bind Parameters
 hybris.fxs.actions.introduce_bind_parameters.description=Replace all literal values in the query with named bind parameters derived from column references
 hybris.fxs.actions.introduce_bind_parameters.no_literals=No literal values found to replace with bind parameters
