+++ modules/shared/core/src/sap/commerce/toolset/HybrisIcons.kt
@@ -244,6 +244,7 @@ object HybrisIcons {
         object Actions {
             val GROUP = getIcon("/icons/flexibleSearch/actions/group.svg")
             val INTRODUCE_PARAMETERS = AllIcons.Actions.InlayRenameInCommentsActive
+            val EXPORT = AllIcons.General.Export
         }
     }
