+++ modules/shared/transform/resources/META-INF/sap.commerce.toolset-shared-transform.xml
@@ -15,11 +15,12 @@
   ~ You should have received a copy of the GNU Lesser General Public License
   ~ along with this program. If not, see <http://www.gnu.org/licenses/>.
   -->
-
 <idea-plugin>
 
-    <extensions defaultExtensionNs="com.intellij">
-        <mcpServer.mcpToolset implementation="sap.commerce.toolset.flexibleSearch.mcp.FlexibleSearchMcpToolset"/>
-    </extensions>
+    <extensionPoints>
+        <extensionPoint qualifiedName="sap.commerce.toolset.transformer"
+                        interface="sap.commerce.toolset.transform.Transformer"
+                        dynamic="true"/>
+    </extensionPoints>
 
 </idea-plugin>
